fix(core): A2's age arrived truncated, so a crit read warn for a second
`classifyStuck` compares the oldest offender's age against `staleSeconds * 4`,
but the query cast it `::int` first. A job 2400.6s past a 600s threshold
therefore reported `warn` until the next whole second ticked over — the one
window in which A2 is escalating is the one where it understated itself.
`::float8` hands the real elapsed time through; the entity label already
rounds to minutes, so nothing downstream changes.

Review flagged this twice as non-blocking. It is one cast, so it is fixed
here rather than carried a sixth round.

+++ packages/core/src/admin/alert-queries.ts
@@ -183,9 +183,10 @@ type StuckRow = {
 
 /** A2 — jobs dispatched or running past staleSeconds (AC 5: BOTH statuses, not dispatched alone). */
 async function alertStuckJobs(staleSeconds: number): Promise<AdminAlert> {
+  // cm:guard age_seconds is float8, never ::int — classifyStuck compares it against staleSeconds * 4, so truncating SQL-side makes a job already past the crit boundary report warn until the next whole second ticks over
   const rows = await db.execute<StuckRow & { total: number }>(sql`
     SELECT j.id, j.type AS job_type, j.dispatched_at,
-           extract(epoch FROM (now() - j.dispatched_at))::int AS age_seconds,
+           extract(epoch FROM (now() - j.dispatched_at))::float8 AS age_seconds,
            count(*) OVER ()::int AS total
     FROM jobs j
     WHERE j.status IN ('dispatched', 'running')
